fix the query over course skills & refactor `ItemFormSection`
Did some further minor adjustements here and there
